+++ src/main/java/roomescape/domain/controller/ReservationController.java
@@ -1,12 +1,12 @@
-package roomescape.controller;
+package roomescape.domain.controller;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import roomescape.service.ReservationService;
-import roomescape.dto.ReservationRequest;
-import roomescape.dto.ReservationResponse;
+import roomescape.domain.service.ReservationService;
+import roomescape.domain.dto.ReservationRequest;
+import roomescape.domain.dto.ReservationResponse;
 
 import java.util.List;
