+++ src/main/java/roomescape/domain/controller/ReservationTimeController.java
@@ -1,12 +1,12 @@
-package roomescape.controller;
+package roomescape.domain.controller;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import roomescape.dto.ReservationTimeRequest;
-import roomescape.dto.ReservationTimeResponse;
-import roomescape.service.ReservationTimeService;
+import roomescape.domain.dto.ReservationTimeRequest;
+import roomescape.domain.dto.ReservationTimeResponse;
+import roomescape.domain.service.ReservationTimeService;
 
 import java.util.List;
