+++ src/main/java/roomescape/domain/dto/ReservationTimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.domain.dto;
 
 import jakarta.validation.constraints.NotBlank;
