+++ src/main/java/roomescape/domain/entity/Reservation.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.entity;
 
 import lombok.Builder;
 import lombok.Getter;
