+++ src/main/java/roomescape/domain/service/ReservationService.java
@@ -1,15 +1,15 @@
-package roomescape.service;
+package roomescape.domain.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.domain.ReservationTime;
-import roomescape.dto.ReservationTimeResponse;
-import roomescape.repository.ReservationRepository;
-import roomescape.domain.Reservation;
-import roomescape.dto.ReservationRequest;
-import roomescape.dto.ReservationResponse;
-import roomescape.repository.ReservationTimeRepository;
+import roomescape.domain.entity.ReservationTime;
+import roomescape.domain.dto.ReservationTimeResponse;
+import roomescape.domain.repository.ReservationRepository;
+import roomescape.domain.entity.Reservation;
+import roomescape.domain.dto.ReservationRequest;
+import roomescape.domain.dto.ReservationResponse;
+import roomescape.domain.repository.ReservationTimeRepository;
 import roomescape.util.DateAndTimeConverter;
 
 import java.util.List;
