+++ src/test/java/roomescape/MissionStepTest.java
@@ -7,8 +7,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.controller.ReservationController;
-import roomescape.domain.Reservation;
+import roomescape.domain.controller.ReservationController;
+import roomescape.domain.entity.Reservation;
 
 import java.lang.reflect.Field;
 import java.sql.Connection;
